rename/document EmptyVec -> VecAllocation
