+++ lib/rust/prelude/src/data.rs
@@ -6,13 +6,11 @@
 // ==============
 
 pub mod at_least_one_of_two;
-pub mod empty_vec;
 pub mod monoid;
 pub mod non_empty_vec;
 pub mod semigroup;
 
 pub use at_least_one_of_two::*;
-pub use empty_vec::*;
 pub use monoid::*;
 pub use non_empty_vec::*;
 pub use semigroup::*;
